requirements.txt
langchain
sentence-transformers
faiss-cpu
rank-bm25
numpy
torch
